requirements.txt
decord==0.6.0
easydict==1.9
einops==0.8.1
flash_attn==2.7.4.post1
func_timeout==4.3.5
horovod==0.28.1
ipdb==0.13.11
kornia==0.6.10
mmengine==0.10.3
nltk==3.8.1
numpy==1.25.2
omegaconf==2.2.3
opencv_python==4.5.4.58
pandas==1.5.1
petrel_oss_sdk==2.2.2.post10086
Pillow==9.3.0
Pillow==11.1.0
PyYAML==6.0
PyYAML==6.0.2
scikit_learn==1.2.2
submitit==1.5.2
timm==0.5.4
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
